core: segmentedControl
