+++ static/app/components/core/segmentedControl/segmentedControl.spec.tsx
